requirements.txt
Django>=5.1.7,<6.0
psycopg2-binary>=2.9.10,<3.0
Pillow>=10.0.0,<11.0
qrcode>=7.4.2,<8.0
python-barcode>=0.16.1,<1.0
djangorestframework>=3.15.1,<4.0
django-cors-headers>=4.3.1,<5.0
django-filter>=25.2,<26.0
gunicorn>=22.0.0,<23.0
whitenoise>=6.6.0,<7.0
dj-database-url>=2.0.0,<3.0
python-decouple>=3.8,<4.0